Rename Cardano.Wallet.StakePool.Metrics to Cardano.Pool.Metrics
