+++ lib/core/cardano-wallet-core.cabal
@@ -89,6 +89,7 @@ library
       Cardano.Pool.DB.Sqlite
       Cardano.Pool.DB.Sqlite.TH
       Cardano.Pool.DB.Sqlite.Types
+      Cardano.Pool.Metrics
       Cardano.Wallet
       Cardano.Wallet.Api
       Cardano.Wallet.Api.Server
@@ -116,7 +117,6 @@ library
       Cardano.Wallet.Primitive.Model
       Cardano.Wallet.Primitive.Types
       Cardano.Wallet.Registry
-      Cardano.Wallet.StakePool.Metrics
       Cardano.Wallet.Transaction
       Cardano.Wallet.Unsafe
       Cardano.Wallet.Version
@@ -202,6 +202,7 @@ test-suite unit
       Cardano.Pool.DB.Arbitrary
       Cardano.Pool.DB.Properties
       Cardano.Pool.DB.SqliteSpec
+      Cardano.Pool.MetricsSpec
       Cardano.Wallet.Api.TypesSpec
       Cardano.Wallet.ApiSpec
       Cardano.Wallet.DB.Arbitrary
@@ -221,7 +222,6 @@ test-suite unit
       Cardano.Wallet.Primitive.CoinSelectionSpec
       Cardano.Wallet.Primitive.FeeSpec
       Cardano.Wallet.Primitive.MnemonicSpec
-      Cardano.Wallet.StakePool.MetricsSpec
       Cardano.Wallet.Primitive.ModelSpec
       Cardano.Wallet.Primitive.TypesSpec
       Cardano.Wallet.RegistrySpec
